+++ confmap/provider/internal/provider_test.go
@@ -28,7 +28,7 @@ import (
 func TestNewRetrievedFromYAML(t *testing.T) {
 	ret, err := NewRetrievedFromYAML([]byte{})
 	require.NoError(t, err)
-	retMap, err := ret.AsMap()
+	retMap, err := ret.AsConf()
 	require.NoError(t, err)
 	assert.Equal(t, confmap.New(), retMap)
 	assert.NoError(t, ret.Close(context.Background()))
@@ -38,7 +38,7 @@ func TestNewRetrievedFromYAMLWithOptions(t *testing.T) {
 	want := errors.New("my error")
 	ret, err := NewRetrievedFromYAML([]byte{}, confmap.WithRetrievedClose(func(context.Context) error { return want }))
 	require.NoError(t, err)
-	retMap, err := ret.AsMap()
+	retMap, err := ret.AsConf()
 	require.NoError(t, err)
 	assert.Equal(t, confmap.New(), retMap)
 	assert.Equal(t, want, ret.Close(context.Background()))