+++ confmap/provider_test.go
@@ -26,7 +26,7 @@ import (
 func TestNewRetrieved(t *testing.T) {
 	ret, err := NewRetrieved(nil)
 	require.NoError(t, err)
-	retMap, err := ret.AsMap()
+	retMap, err := ret.AsConf()
 	require.NoError(t, err)
 	assert.Equal(t, New(), retMap)
 	assert.NoError(t, ret.Close(context.Background()))
@@ -36,7 +36,7 @@ func TestNewRetrievedWithOptions(t *testing.T) {
 	want := errors.New("my error")
 	ret, err := NewRetrieved(nil, WithRetrievedClose(func(context.Context) error { return want }))
 	require.NoError(t, err)
-	retMap, err := ret.AsMap()
+	retMap, err := ret.AsConf()
 	require.NoError(t, err)
 	assert.Equal(t, New(), retMap)
 	assert.Equal(t, want, ret.Close(context.Background()))